Added HTTP server and client communication for Domoticz home automation server. WiFi connection is established based on project_defines.h entries (needs to be modified to fit current network setup. Connection to WiFi has recconection mechanism in case of lost connection.
HTTP server is used to control the light switch, while HTTP client is
used to update json for Domoticz in case of changing light switch state
by an external source.

All unnecessary (for home automation project) code was removed from
sketch file.

HTTPClient is currently added to ControlModule, might move it to some
external class. It's not called from inside the class, so it's not
needed as a part of it.